refactor: Installer
