+++ src-tauri/src/core/language/python.rs
@@ -10,10 +10,8 @@ use crate::core::utils::config::{del_language, versions_list};
 use async_trait::async_trait;
 use lvm_core::config::get::{get_config_bool, get_language_current_version};
 use lvm_core::enums::path::EPath;
-use lvm_core::files::get::get_dirs;
 use lvm_core::path::get::current_path;
 use std::fs;
-use std::io::ErrorKind;
 use std::path::PathBuf;
 use tauri::Wry;
 
@@ -61,19 +59,6 @@ impl PythonInstaller {
             "unknown".to_string()
         }
     }
-
-    fn name(&self) -> &'static str {
-        "python"
-    }
-
-    // Get base directory
-    fn get_base_dir(&self) -> PathBuf {
-        let path = EPath::Version.path().join(self.name());
-        if !path.exists() {
-            fs::create_dir_all(&path).expect("create err");
-        }
-        path
-    }
 }
 
 #[async_trait]
@@ -82,21 +67,6 @@ impl LanguageInstaller for PythonInstaller {
         let versions = versions_list("python", fetch_versions_python).await?;
 
         Ok(versions)
-    }
-
-    async fn list_installed(&self) -> Result<Vec<String>, String> {
-        let dir = self.get_base_dir();
-        get_dirs(&dir).map_err(|e| e.to_string())
-    }
-
-    async fn current(&self) -> Result<Option<String>, String> {
-        let path = self.get_base_dir().join("current");
-
-        match fs::read_to_string(path) {
-            Ok(v) => Ok(Some(v.trim().to_string())),
-            Err(e) if e.kind() == ErrorKind::NotFound => Ok(None),
-            Err(e) => Err(e.to_string()),
-        }
     }
 
     async fn install(
@@ -147,12 +117,8 @@ impl LanguageInstaller for PythonInstaller {
         Ok(())
     }
 
-    async fn activate(&self, version: &str) -> Result<(), String> {
-        let current_file = current_path(self.name());
-
-        fs::write(current_file, version).map_err(|e| e.to_string())?;
-
-        Ok(())
+    fn name(&self) -> &str {
+        "python"
     }
 
     async fn deactivate(&self, version: &str) -> Result<(), String> {
